Send UEP startup failure AMQP message
When an endpoint fails to start up, if the endpoint manager has at least
acquired the UEP's AMQP credentials, it can send a failure status to the web
services.  In the majority of cases then, the SDK need not wait for the
web-service to time out -- it can get near-instantaneous feedback.

This changeset addresses both the pre-exec states (e.g., the user cannot be
mapped to local uid) and in the post-exec UEP startup failures (e.g., an
invalid configuration).

[sc-28883]